Gmail-MCP/pyproject.toml
[project]
name = "gmail-mcp"
version = "0.1.0"
description = "Gmail MCP authentication and API project"
dependencies = [
    "python-dotenv",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "fastmcp",
    "typing-extensions"
]
